W174f: mdlFile truncation scare was a FALSE ALARM; fclShopMisc clears four more windows.
INCIDENT: Z03MdlC reported mdlFile.c collapsing 55,717 -> 16,443 lines and correctly halted the file. Investigation shows NO data loss -- the file is 55,710 lines / 1,043,810 bytes / **667 unique markers**, versus 55,716 / 1,043,880 / 667 committed; the 70-byte delta is Z04's removal of stray duplicate lines after 0034B6A0. Zero markers lost, zero COMPILE_ERROR, zero STALE, zero lost MATCHes, no new over-window. The short read was a transient snapshot taken while a peer's write was in flight, not a truncation. Z03's call to halt was still the right one: with ~25 agents on one tree, the cost of a false halt is minutes, the cost of a missed real truncation is the whole wave.

All claimed work survives and is re-verified by content: 0033F6A0 MATCH (allocator _u32 -> _abs, param1 u32 -> int swept through prototype and definition, redundant casts removed, param2 branch inverted to the matching sibling's order), 00349870 and 0034A650 MATCH, and the whole 00349AF0/00349FD0/0034A370/0034A9C0/0034AE30 family MATCH. Also 00356B00 (160/160) and 00355DD0 (156/160) MATCH.

fclShopMisc FOUR more window clears, over-window 8 -> 4, every one by deleting invented work:
  003F4A90  900/848 nd586 -> 848/848 nd582   signed integer result recovered; UNSIGNED float-to-int conversion expansion removed
  003F55B0  644/640 nd329 -> 632/640 nd317   byte semantic type at the draw call kills an invented masked-colour local
  003F7D50  660/656 nd456 -> 644/656 nd190   invented LONG PROMOTIONS removed from two int/short comparisons
  003FBC00  628/624 nd341 -> 624/624 nd362   nmlist lookup and text draw routed through existing typed pointer/f32 aliases

STANDING RULE REINFORCED: never write a whole file back from an in-memory copy. One agent's edit reported 'recovered from a stale file hash using a previous read snapshot' -- that recovery path is the likely source of both the transient short read and the twice-reintroduced duplicate EOF lines. Splice only marker-bounded regions, re-reading from disk immediately before every write.

+++ src/Camp/h_camp.c
@@ -1693,11 +1693,11 @@ void h_campUpdateRootMenuEntryFadeOut(CampRootDrawWork* work, f32 alpha)
     func_00115980(particle);
 
     slide = (200.0f * (f32)(0x16 - work->frame)) / 22.0f;
-    oldDuration = (f32)work->transitionDuration;
-    work->transitionDuration = (u32)(oldDuration + 1.0f);
-    newDuration = (f32)work->transitionDuration;
+    oldDuration = (f32)(s32)work->transitionDuration;
+    work->transitionDuration = (s32)(oldDuration + 1.0f);
+    newDuration = (f32)(s32)work->transitionDuration;
     if (!(newDuration <= 767.0f)) {
-        work->transitionDuration = (u32)(newDuration - 448.0f);
+        work->transitionDuration = (s32)(newDuration - 448.0f);
     }
 
     x = 589.0f + slide;
@@ -1808,12 +1808,12 @@ void h_campUpdateRootMenuEntryFinish(CampRootDrawWork* work, f32 alpha)
     f32 temp_f22;
 
     sp48 = 0;
-    temp_f1 = (f32)work->transitionDuration;
+    temp_f1 = (f32)(s32)work->transitionDuration;
     sp44 = temp_f1;
-    work->transitionDuration = (u32)(temp_f1 + 1.0f);
-    temp_f1_2 = (f32)work->transitionDuration;
+    work->transitionDuration = (s32)(temp_f1 + 1.0f);
+    temp_f1_2 = (f32)(s32)work->transitionDuration;
     if (!(temp_f1_2 <= (f32)0x2ff)) {
-        work->transitionDuration = (u32)(temp_f1_2 - 448.0f);
+        work->transitionDuration = (s32)(temp_f1_2 - 448.0f);
     }
     sp40 = *(s64*)&sp48;
     temp_f21 = 589.0f + *(f32*)&sp40;
@@ -1881,11 +1881,11 @@ void h_campUpdateRootMenuSelectionEffect(CampRootDrawWork* work, f32 alpha)
         work->drawChild = (KwlnTask*)1;
     }
 
-    oldDuration = (f32)work->transitionDuration;
-    work->transitionDuration = (u32)(oldDuration + 1.0f);
-    newDuration = (f32)work->transitionDuration;
+    oldDuration = (f32)(s32)work->transitionDuration;
+    work->transitionDuration = (s32)(oldDuration + 1.0f);
+    newDuration = (f32)(s32)work->transitionDuration;
     if (!(newDuration <= 767.0f)) {
-        work->transitionDuration = (u32)(newDuration - 448.0f);
+        work->transitionDuration = (s32)(newDuration - 448.0f);
     }
 
     pos.x = 0.0f;
